+++ test/file_io_test.py
@@ -5,7 +5,7 @@ import pytest
 from grpclib import Status
 from grpclib.exceptions import GRPCError
 
-from modal.file_io import FileIO, FileWatchEvent, WatchEvent, delete_bytes, replace_bytes
+from modal.file_io import FileIO, FileWatchEvent, FileWatchEventType, delete_bytes, replace_bytes
 from modal_proto import api_pb2
 
 OPEN_EXEC_ID = "exec-open-123"
@@ -19,6 +19,9 @@ WRITE_REPLACE_EXEC_ID = "exec-write-replace-123"
 DELETE_EXEC_ID = "exec-delete-123"
 CLOSE_EXEC_ID = "exec-close-123"
 WATCH_EXEC_ID = "exec-watch-123"
+LS_EXEC_ID = "exec-ls-123"
+MKDIR_EXEC_ID = "exec-mkdir-123"
+RM_EXEC_ID = "exec-rm-123"
 
 
 async def container_filesystem_exec(servicer, stream):
@@ -74,6 +77,12 @@ async def container_filesystem_exec(servicer, stream):
         await stream.send_message(api_pb2.ContainerFilesystemExecResponse(exec_id=CLOSE_EXEC_ID))
     elif req.HasField("file_watch_request"):
         await stream.send_message(api_pb2.ContainerFilesystemExecResponse(exec_id=WATCH_EXEC_ID))
+    elif req.HasField("file_ls_request"):
+        await stream.send_message(api_pb2.ContainerFilesystemExecResponse(exec_id=LS_EXEC_ID))
+    elif req.HasField("file_mkdir_request"):
+        await stream.send_message(api_pb2.ContainerFilesystemExecResponse(exec_id=MKDIR_EXEC_ID))
+    elif req.HasField("file_rm_request"):
+        await stream.send_message(api_pb2.ContainerFilesystemExecResponse(exec_id=RM_EXEC_ID))
 
 
 def test_file_read(servicer, client):
@@ -350,9 +359,9 @@ def test_client_retry(servicer, client):
 def test_file_watch(servicer, client):
     """Test file watching."""
     expected_events = [
-        FileWatchEvent(paths=["/foo.txt"], event=WatchEvent.Access),
-        FileWatchEvent(paths=["/bar.txt"], event=WatchEvent.Create),
-        FileWatchEvent(paths=["/baz.txt", "/baz/foo.txt"], event=WatchEvent.Modify),
+        FileWatchEvent(paths=["/foo.txt"], type=FileWatchEventType.Access),
+        FileWatchEvent(paths=["/bar.txt"], type=FileWatchEventType.Create),
+        FileWatchEvent(paths=["/baz.txt", "/baz/foo.txt"], type=FileWatchEventType.Modify),
     ]
 
     async def container_filesystem_exec_get_output(servicer, stream):
@@ -366,14 +375,6 @@ def test_file_watch(servicer, client):
                         ]
                     )
                 )
-            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
-        else:
-            raise Exception("Unexpected exec_id: " + req.exec_id)
-
-    with servicer.intercept() as ctx:
-        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
-        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
-
         events = FileIO.watch("/test.txt", client, "task-123")
         seen_events: list[FileWatchEvent] = []
         for event in events:
@@ -391,9 +392,7 @@ def test_file_watch_ignore_invalid_events(servicer, client):
         req = await stream.recv_message()
         if req.exec_id == WATCH_EXEC_ID:
             await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(output=[b"invalid\n\n"]))
-            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
-        else:
-            raise Exception("Unexpected exec_id: " + req.exec_id)
+        await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
 
     with servicer.intercept() as ctx:
         ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
@@ -404,3 +403,97 @@ def test_file_watch_ignore_invalid_events(servicer, client):
         for event in events:
             seen_events.append(event)
         assert len(seen_events) == 0
+
+
+@pytest.mark.asyncio
+async def test_file_io_async_context_manager(servicer, client):
+    """Test file io context manager."""
+    content = "foo\nbar\nbaz\n"
+
+    async def container_filesystem_exec_get_output(servicer, stream):
+        req = await stream.recv_message()
+        if req.exec_id == READ_EXEC_ID:
+            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(output=[content.encode()]))
+        await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
+
+    with servicer.intercept() as ctx:
+        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
+        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
+
+        async with FileIO.create("/test.txt", "w+", client, "task-123") as f:
+            await f.write.aio(content)
+            assert await f.read.aio() == content
+
+
+def test_file_io_sync_context_manager(servicer, client):
+    """Test file io context manager."""
+    content = "foo\nbar\nbaz\n"
+
+    async def container_filesystem_exec_get_output(servicer, stream):
+        req = await stream.recv_message()
+        if req.exec_id == READ_EXEC_ID:
+            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(output=[content.encode()]))
+        await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
+
+    with servicer.intercept() as ctx:
+        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
+        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
+
+        with FileIO.create("/test.txt", "w+", client, "task-123") as f:
+            f.write(content)
+            assert f.read() == content
+
+
+def test_ls(servicer, client):
+    """Test ls."""
+
+    async def container_filesystem_exec_get_output(servicer, stream):
+        req = await stream.recv_message()
+        if req.exec_id == LS_EXEC_ID:
+            await stream.send_message(
+                api_pb2.FilesystemRuntimeOutputBatch(output=[b'{"paths": ["foo", "bar", "baz"]}'])
+            )
+            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
+        else:
+            raise Exception("Unexpected exec_id: " + req.exec_id)
+
+    with servicer.intercept() as ctx:
+        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
+        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
+
+        files = FileIO.ls("/test.txt", client, "task-123")
+        assert files == ["foo", "bar", "baz"]
+
+
+def test_mkdir(servicer, client):
+    """Test mkdir."""
+
+    async def container_filesystem_exec_get_output(servicer, stream):
+        req = await stream.recv_message()
+        if req.exec_id == MKDIR_EXEC_ID:
+            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
+        else:
+            raise Exception("Unexpected exec_id: " + req.exec_id)
+
+    with servicer.intercept() as ctx:
+        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
+        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
+
+        FileIO.mkdir("/test.txt", client, "task-123")
+
+
+def test_rm(servicer, client):
+    """Test rm."""
+
+    async def container_filesystem_exec_get_output(servicer, stream):
+        req = await stream.recv_message()
+        if req.exec_id == RM_EXEC_ID:
+            await stream.send_message(api_pb2.FilesystemRuntimeOutputBatch(eof=True))
+        else:
+            raise Exception("Unexpected exec_id: " + req.exec_id)
+
+    with servicer.intercept() as ctx:
+        ctx.set_responder("ContainerFilesystemExec", container_filesystem_exec)
+        ctx.set_responder("ContainerFilesystemExecGetOutput", container_filesystem_exec_get_output)
+
+        FileIO.rm("/test.txt", client, "task-123")